返ってきた値をそのまま渡せるよう、日本語ではなく英語を返す
